% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{.evaluation_queue_query}
\alias{.evaluation_queue_query}
\title{Wraps challengeutils.utils.evaluation_queue_query. Written for tests}
\usage{
.evaluation_queue_query(query)
}
\arguments{
\item{query}{A URI for evaluation queues (select * from evaluation_12345)}

\item{syn}{A Synapse object}
}
\description{
Wraps challengeutils.utils.evaluation_queue_query. Written for tests
}
