% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation_queue_query}
\alias{evaluation_queue_query}
\title{Queries an evaluation queue for submissions}
\usage{
evaluation_queue_query(syn, query)
}
\arguments{
\item{syn}{A Synapse object}

\item{uri}{A URI for evaluation queues (select * from evaluation_12345)}
}
\value{
The leaderboard results in a tibble.
}
\description{
Queries an evaluation queue for submissions
}
\examples{
library(challengerutils)
use_condaenv('challenge')
synapseclient <- import('synapseclient')
syn <- synapseclient$Synapse()
syn$login()
leaderboard_results <- evaluation_queue_query(syn, 'select * from evaluation_12345')
}
