% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{syn_login}
\alias{syn_login}
\title{Logs into Synapse.}
\usage{
syn_login(username = NULL, password = NULL)
}
\arguments{
\item{username}{a Synapse username (optional, not required if .synapseConfig available)}

\item{password}{a Synapse password (optional, not required if .synapseConfig available)}
}
\description{
Logs into Synapse.
}
\examples{
library(challengerutils)
syn_login()
}
