% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation_queue_query}
\alias{evaluation_queue_query}
\title{Queries an evaluation queue for submissions}
\usage{
evaluation_queue_query(query)
}
\arguments{
\item{query}{A URI for evaluation queues (select * from evaluation_12345)}
}
\value{
The leaderboard results in a tibble.
}
\description{
Queries an evaluation queue for submissions
}
\examples{
library(challengerutils)
use_condaenv('challenge')
syn_login()
leaderboard_results <- evaluation_queue_query('select * from evaluation_9614194')
}
